Move all behavior related to an agent into it's own folder
