+++ agent/cache.py
