+++ agent/data/prompts/objective/examples/response.yaml
