+++ agent/data/prompts/objective/examples/thought.yaml
