+++ agent/data/prompts/objective/human/thought.yaml
