+++ agent/data/prompts/objective/summaries/thought.yaml
