+++ agent/data/prompts/objective/system/thought.yaml
